# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Robot Rangers",
        "description": "Lista de URI's de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://robot-rangers-laravel-git-crud-api-robotrangers.vercel.app/"
        }
    ],
    "paths": {
...
        "/rest/marcas/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Retorna la información de una marca especifica",
                "operationId": "0b6865c623deb71fd8abbc07d7584b88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-05-07 19:57:30"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-05-07 19:57:30"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nike"
                                        },
                                        "imagen": {
                                            "type": "string",
                                            "example": "https://static.ffx.io/images/$zoom_1%2C$multiply_2.1113%2C$ratio_1.777778%2C$width_485%2C$x_53%2C$y_54/t_crop_custom/q_86%2Cf_auto/f54a63f81d65d0b5b4db5f2235bbe665f5310227"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Just do it"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Marca] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}